New feature on admin-cart , and improved sample index, and audit report

+++ services/web/app/storage/api/misc.py
@@ -39,6 +39,9 @@ from ...storage.enums import FixedColdStorageType, FixedColdStorageTemps
 @api.route("/storage/transfer/rack_to_shelf", methods=["POST"])
 @token_required
 def storage_transfer_rack_to_shelf(tokenuser: UserAccount):
+    """Only add new entitytostorage record for rack to shelf,
+    update/delete is done by transfer rack to cart or editing rack info"""
+
     values = request.get_json()
 
     if not values:
@@ -49,15 +52,30 @@ def storage_transfer_rack_to_shelf(tokenuser: UserAccount):
     except ValidationError as err:
         return validation_error_response(err)
 
-    ets = EntityToStorage.query.filter_by(
-        rack_id=values["rack_id"], storage_type="BTS"
-    ).first()
+    # ets = EntityToStorage.query.filter_by(
+    #    rack_id=values["rack_id"], storage_type="BTS"
+    # ).first()
+    etss = (
+        EntityToStorage.query.filter_by(rack_id=values["rack_id"], storage_type="BTS")
+        .order_by(EntityToStorage.removed.asc())
+        .all()
+    )
 
-    if ets is not None:
-        ets.box_id = None
-        ets.shelf_id = values["shelf_id"]
-        ets.editor_id = tokenuser.id
-        ets.storage_type = "BTS"
+    if len(etss) > 0:
+        n = 0
+        for ets in etss:
+            n = n + 1
+            if n == 1:
+                # ets.rack_id = values["rack_id"]
+                ets.shelf_id = values["shelf_id"]
+                ets.storage_type = "BTS"
+                ets.removed = False
+                ets.update({"editor_id": tokenuser.id})
+                db.session.add(ets)
+            else:
+
+                ets.update({"editor_id": tokenuser.id})
+                db.session.delete(ets)
 
     else:
         ets = EntityToStorage(**rack_to_shelf_result)
@@ -76,6 +94,8 @@ def storage_transfer_rack_to_shelf(tokenuser: UserAccount):
 @api.route("/storage/transfer/racks_to_shelf", methods=["POST"])
 @token_required
 def storage_transfer_racks_to_shelf(tokenuser: UserAccount):
+    """Only add new entitytostorage record for rack to shelf,
+    update/delete is done by transfer rack to cart or editing rack info"""
     values = request.get_json()
 
     if not values:
@@ -95,31 +115,42 @@ def storage_transfer_racks_to_shelf(tokenuser: UserAccount):
         except ValidationError as err:
             return validation_error_response(err)
 
-        # etss = EntityToStorage.query.filter_by(rack_id=values["rack_id"],
-        #             storage_type='BTS', removed=True).all()
-        # if len(etss)>0:
-        #     try:
-        #         for ets in etss:
-        #             # ets.removed = True
-        #             # ets.editor_id = tokenuser.id
-        #             # ets.updated_on = func.now()
-        #             # db.session.add(ets)
-        #             db.session.delete(ets)
-        #
-        #         db.session.flush()
-        #         print('okok')
-        #     except Exception as err:
-        #         return transaction_error_response(err)
+        etss = (
+            EntityToStorage.query.filter_by(
+                rack_id=values["rack_id"], storage_type="BTS"
+            )
+            .order_by(EntityToStorage.removed)
+            .all()
+        )
+        if len(etss) > 0:
 
-        ets = EntityToStorage(**rack_to_shelf_result)
-        ets.author_id = tokenuser.id
-        ets.storage_type = "BTS"
-        try:
+            try:
+                n = 0
+                for ets in etss:
+                    n = n + 1
+                    if n > 1:
+                        ets.update({"editor_id": tokenuser.id})
+                        db.session.delete(ets)
+                    else:
+                        ets.removed = False
+                        ets.shelf_id = rack_values["shelf_id"]
+                        ets.update({"editor_id": tokenuser.id})
+                        db.session.add(ets)
+
+                db.session.flush()
+            except Exception as err:
+                return transaction_error_response(err)
+        else:
+            ets = EntityToStorage(**rack_to_shelf_result)
+            ets.author_id = tokenuser.id
+            ets.storage_type = "BTS"
             db.session.add(ets)
-            db.session.flush()
 
-        except Exception as err:
-            return transaction_error_response(err)
+            try:
+                db.session.flush()
+
+            except Exception as err:
+                return transaction_error_response(err)
 
         ucs = UserCart.query.filter_by(rack_id=rack_id, storage_type="RUC").all()
 
@@ -151,6 +182,7 @@ def storage_transfer_racks_to_shelf(tokenuser: UserAccount):
 
             try:
                 db.session.add(rk)
+                db.session.flush()
             except Exception as err:
                 return transaction_error_response(err)
 
@@ -203,37 +235,48 @@ def storage_transfer_sample_to_shelf(tokenuser: UserAccount):
     except ValidationError as err:
         return validation_error_response(err)
 
-    etss = EntityToStorage.query.filter(sample_id=values["sample_id"]).all()
+    etss = (
+        EntityToStorage.query.filter(sample_id=values["sample_id"])
+        .order_by(EntityToStorage.removed)
+        .all()
+    )
     if len(etss) > 0:
         # warning, confirmation
         try:
+            n = 0
             for ets in etss:
-                db.session.delete(ets)
+                n = n + 1
+                if n > 1:
+                    ets.update({"editor_i": tokenuser.id})
+                    db.session.delete(ets)
+                    continue
+                else:
+                    ets.rack_id = None
+                    ets.shelf_id = values["shelf_id"]
+                    ets.storage_type = "STS"
+                    ets.removed = False
+                    ets.update({"editor_id": tokenuser.id})
+                    db.session.add(ets)
+
+            db.session.flush()
         except Exception as err:
             return transaction_error_response(err)
 
-    #
-    # etss = EntityToStorage.query.filter_by(sample_id=values["sample_id"], storage_type='STS').all()
-    # if ets != None:
-    #     ets.box_id = None
-    #     ets.shelf_id = values["shelf_id"]
-    #     ets.editor_id = tokenuser.id
-    #     ets.updated_on = func.now()
-    #     ets.storage_type = "STS"
-
-    ets = EntityToStorage(
-        sample_id=values["sample_id"],
-        shelf_id=values["shelf_id"],
-        storage_type="STS",
-        entry=values["entry"],
-        entry_datetime=values["entry_datetime"],
-        author_id=tokenuser.id,
-    )
-    try:
+    else:
+        ets = EntityToStorage(
+            sample_id=values["sample_id"],
+            shelf_id=values["shelf_id"],
+            storage_type="STS",
+            entry=values["entry"],
+            entry_datetime=values["entry_datetime"],
+            author_id=tokenuser.id,
+        )
         db.session.add(ets)
-        db.session.flush()
-    except Exception as err:
-        return transaction_error_response(err)
+
+        try:
+            db.session.flush()
+        except Exception as err:
+            return transaction_error_response(err)
 
     usercart = UserCart.query.filter_by(
         sample_id=values["sample_id"], author_id=tokenuser.id
@@ -271,37 +314,54 @@ def storage_transfer_samples_to_shelf(tokenuser: UserAccount):
         except ValidationError as err:
             return validation_error_response(err)
 
-        # etss = EntityToStorage.query.filter_by(sample_id=values["sample_id"], storage_type='STB').all()
-        etss = EntityToStorage.query.filter_by(sample_id=values["sample_id"]).all()
-        # TODO: instead of deleting exisint etss add new ets instance and set the existing ets to: removed=True
+        etss = (
+            EntityToStorage.query.filter_by(sample_id=values["sample_id"])
+            .order_by(EntityToStorage.removed)
+            .all()
+        )
+
         if len(etss) > 0:
-            # warning, confirmation
+            # Update existing entitytostorage(ets) record for the given sample
+            # Only keep one ets record for one sample
             try:
+                n = 0
                 for ets in etss:
-                    db.session.delete(ets)
+                    n = n + 1
+                    if n == 1:
+                        ets.rack_id = None
+                        ets.col = None
+                        ets.row = None
+                        ets.shelf_id = values["shelf_id"]
+                        ets.storage_type = "STS"
+                        ets.entry = values["entry"]
+                        ets.entry_datetime = values["entry_datetime"]
+                        ets.removed = False
+                        ets.update({"editor_id": tokenuser.id})
+                        db.session.add(ets)
+                    else:
+                        ets.update({"editor_id": tokenuser.id})
+                        db.session.delete(ets)
+                        continue
+
+                db.session.flush()
             except Exception as err:
                 return transaction_error_response(err)
-        # ets = EntityToStorage.query.filter_by(sample_id=values["sample_id"], storage_type='STS').first()
-        # if ets != None:
-        #     ets.box_id = None
-        #     ets.shelf_id = values["shelf_id"]
-        #     ets.editor_id = tokenuser.id
-        #     ets.updated_on = func.now()
-        #     ets.storage_type = "STS"
 
-        ets = EntityToStorage(
-            sample_id=values["sample_id"],
-            shelf_id=values["shelf_id"],
-            storage_type="STS",
-            entry=values["entry"],
-            entry_datetime=values["entry_datetime"],
-            author_id=tokenuser.id,
-        )
-        try:
-            db.session.add(ets)
-            db.session.flush()
-        except Exception as err:
-            return transaction_error_response(err)
+        else:
+            # Add new ets record for the given sample
+            ets = EntityToStorage(
+                sample_id=values["sample_id"],
+                shelf_id=values["shelf_id"],
+                storage_type="STS",
+                entry=values["entry"],
+                entry_datetime=values["entry_datetime"],
+                author_id=tokenuser.id,
+            )
+            try:
+                db.session.add(ets)
+                db.session.flush()
+            except Exception as err:
+                return transaction_error_response(err)
 
         usercart = UserCart.query.filter_by(
             sample_id=values["sample_id"], author_id=tokenuser.id
@@ -440,12 +500,13 @@ def storage_view_panel(tokenuser: UserAccount):
 def storage_shelf_overview(tokenuser: UserAccount):
 
     sites = [tokenuser.site_id]
-    try:
-        choices0 = tokenuser.settings["data_entry"]["site"]["choices"]
-        if len(choices0) > 0:
-            sites = list(set([sites + choices0]))
-    except:
-        pass
+    if not tokenuser.is_admin:
+        try:
+            choices0 = tokenuser.settings["data_entry"]["site"]["choices"]
+            if len(choices0) > 0:
+                sites = list(set([sites + choices0]))
+        except:
+            pass
 
     locations = (
         db.session.query(SiteInformation)
